Update to_json Indent Type Annotation
Updated the to_json methods indent type annotation on the AWSObject, and
AWSHelperFn to allow for None. Added unit tests to ensure the string
representation is correct.

Sending an integer, or string to 'json.dumps' indent parameter results in a
pretty printed json string with new lines in it. Sending in None results in
the most compact representation.